+++ TextShooter/BIDLevelSceneFactory.h
@@ -11,6 +11,6 @@
 
 @interface BIDLevelSceneFactory : NSObject
 
-+ (Class)sceneClassForLevelNumber:(int)number mode:(BIDGameMode)mode;
++ (Class)sceneClassForLevelNumber:(NSUInteger)number mode:(BIDGameMode)mode;
 
 @end
